feat(gatsby): do not overwrite GraphQL type file if exists and the contents are the same

+++ packages/gatsby/src/utils/graphql-typegen/ts-codegen.ts
@@ -160,5 +160,11 @@ export async function writeTypeScriptTypes(
     ...codegenOptions,
   })
 
+  if (
+    (await fs.pathExists(filename)) &&
+    result === (await fs.readFile(filename, `utf8`))
+  ) {
+    return
+  }
   await fs.outputFile(filename, result)
 }
